ai-service/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning & Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9

# Caching
redis==5.0.1

# Logging
loguru==0.7.2

# Data Validation
marshmallow==3.20.1

# File Processing
python-multipart==0.0.6
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Date & Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
aiofiles==23.2.1

# Code Analysis Tools
pylint==3.0.3
flake8==6.1.0
bandit==1.7.5

# Git Operations
GitPython==3.1.40

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring
prometheus-client==0.19.0

# Security
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Additional ML Libraries
scipy==1.11.4
matplotlib==3.8.2
seaborn==0.13.0

# Text Processing
fuzzy-wuzzy==0.18.0
python-Levenshtein==0.23.0

# API Documentation
fastapi-users==12.1.2

# Async Support
aiofiles==23.2.1
aioredis==2.0.1